pyproject.toml
[project]
name = "jboggs-personal-site"
version = "0.1.0"
description = "Jared Boggs's personal website."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "rasterio>=1.4.3",
    "scipy>=1.16.2",
]
